Replace old notices with print tables
Previous notices were not very clear and made the code more
confusing, replaced them with printing tables that are already
bring created anyway.
Log output is much easier to follow this way
